+++ packages/charlie/CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## unreleased (NaN-NaN-NaN)
 
+### Changed
+
+ - Changed the `waistbandWidth` option type from `mm` to `pct`
+
 ### Fixed
 
  - Support drafting of non-human measurements (dolls & giants) Closes [#1313](https://github.com/freesewing/freesewing/issues/1313)